+++ src/components/index.js
@@ -16,10 +16,8 @@ require('./visible');
 require('./wasd-controls');
 
 require('./scene/canvas');
+require('./scene/debug');
 require('./scene/fog');
 require('./scene/keyboard-shortcuts');
 require('./scene/stats');
 require('./scene/vr-mode-ui');
-
-// Deprecated.
-require('./loader');