+++ tests/components/geometry.test.js
@@ -9,7 +9,7 @@ var degToRad = require('index').THREE.Math.degToRad;
 suite('geometry', function () {
   setup(function (done) {
     var el = this.el = helpers.entityFactory();
-    el.setAttribute('geometry', 'buffer: false; primitive: box');
+    el.setAttribute('geometry', 'buffer: false; primitive: box;');
     el.addEventListener('loaded', function () {
       done();
     });
@@ -55,6 +55,35 @@ suite('geometry', function () {
     });
   });
 
+  suite('merge geometries', function () {
+    setup(function (done) {
+      var self = this;
+      var targetEl = this.targetEl = helpers.entityFactory();
+      targetEl.setAttribute('geometry', 'buffer: false; primitive: box;');
+      targetEl.addEventListener('loaded', function () {
+        var sourceEl = self.sourceEl = document.createElement('a-entity');
+        targetEl.sceneEl.appendChild(sourceEl);
+        sourceEl.addEventListener('loaded', function () {
+          done();
+        });
+      });
+    });
+
+    test('merges geometries', function () {
+      var sourceEl = this.sourceEl;
+      var targetEl = this.targetEl;
+      var sceneEl = sourceEl.sceneEl;
+      var targetGeometry = targetEl.getObject3D('mesh').geometry;
+      targetEl.id = 'mergeTarget';
+      sourceEl.id = 'mergeSource';
+      assert.ok(sceneEl.querySelector('#mergeSource'));
+      assert.equal(targetGeometry.vertices.length, 8);
+      sourceEl.setAttribute('geometry', 'buffer: false; skipCache: true; primitive: box; mergeTo: #mergeTarget');
+      assert.equal(sceneEl.querySelector('#mergeSource'), null);
+      assert.equal(targetGeometry.vertices.length, 16);
+    });
+  });
+
   suite('remove', function () {
     test('removes geometry', function () {
       var mesh = this.el.getObject3D('mesh');
@@ -67,67 +96,6 @@ suite('geometry', function () {
       var disposeSpy = this.sinon.spy(geometry, 'dispose');
       this.el.removeAttribute('geometry');
       assert.ok(disposeSpy.called);
-    });
-  });
-
-  suite('translate', function () {
-    var DEFAULT_VERTICES = [
-      {x: 0.5, y: 0.5, z: 0.5}, {x: 0.5, y: 0.5, z: -0.5}, {x: 0.5, y: -0.5, z: 0.5},
-      {x: 0.5, y: -0.5, z: -0.5}, {x: -0.5, y: 0.5, z: -0.5}, {x: -0.5, y: 0.5, z: 0.5},
-      {x: -0.5, y: -0.5, z: -0.5}, {x: -0.5, y: -0.5, z: 0.5}
-    ];
-
-    setup(function () {
-      this.el.setAttribute('geometry', {
-        buffer: false,
-        primitive: 'box',
-        depth: 1,
-        height: 1,
-        width: 1
-      });
-    });
-
-    test('defaults translate to center', function () {
-      assert.shallowDeepEqual(this.el.getObject3D('mesh').geometry.vertices, DEFAULT_VERTICES);
-    });
-
-    test('can set translate', function () {
-      var el = this.el;
-      el.setAttribute('geometry', 'translate', '-2 4 2');
-      assert.shallowDeepEqual(el.getObject3D('mesh').geometry.vertices, [
-        {x: -1.5, y: 4.5, z: 2.5}, {x: -1.5, y: 4.5, z: 1.5}, {x: -1.5, y: 3.5, z: 2.5},
-        {x: -1.5, y: 3.5, z: 1.5}, {x: -2.5, y: 4.5, z: 1.5}, {x: -2.5, y: 4.5, z: 2.5},
-        {x: -2.5, y: 3.5, z: 1.5}, {x: -2.5, y: 3.5, z: 2.5}]);
-    });
-
-    test('can update translate', function (done) {
-      var el = this.el;
-      el.setAttribute('geometry', 'translate', '-2 4 2');
-      el.setAttribute('geometry', 'translate', '0 0 0');
-      setTimeout(function () {
-        assert.shallowDeepEqual(el.getObject3D('mesh').geometry.vertices, DEFAULT_VERTICES);
-        done();
-      });
-    });
-
-    test('can remove translate', function () {
-      var el = this.el;
-      el.setAttribute('geometry', 'translate', '-2 4 2');
-      this.el.setAttribute('geometry', {
-        buffer: false,
-        primitive: 'box',
-        depth: 1,
-        height: 1,
-        width: 1
-      });
-      assert.shallowDeepEqual(el.getObject3D('mesh').geometry.vertices, DEFAULT_VERTICES);
-    });
-
-    test('does not recreate geometry when just translating', function () {
-      var el = this.el;
-      var uuid = el.getObject3D('mesh').geometry.uuid;
-      el.setAttribute('geometry', 'translate', '-2 4 2');
-      assert.equal(el.getObject3D('mesh').geometry.uuid, uuid);
     });
   });
 
@@ -157,6 +125,7 @@ suite('standard geometries', function () {
       buffer: false, primitive: 'circle', radius: 5, segments: 4, thetaStart: 0,
       thetaLength: 350
     });
+
     geometry = el.getObject3D('mesh').geometry;
     assert.equal(geometry.type, 'CircleGeometry');
     assert.equal(geometry.parameters.radius, 5);
@@ -172,6 +141,7 @@ suite('standard geometries', function () {
       buffer: false, primitive: 'cylinder', radius: 1, height: 2, segmentsRadial: 3,
       segmentsHeight: 4, openEnded: true, thetaStart: 240, thetaLength: 350
     });
+
     geometry = el.getObject3D('mesh').geometry;
     assert.equal(geometry.type, 'CylinderGeometry');
     assert.equal(geometry.parameters.radiusTop, 1);
@@ -191,6 +161,7 @@ suite('standard geometries', function () {
       buffer: false, primitive: 'cone', radiusTop: 1, radiusBottom: 5, height: 2,
       segmentsRadial: 3, segmentsHeight: 4, openEnded: true, thetaStart: 240, thetaLength: 350
     });
+
     geometry = el.getObject3D('mesh').geometry;
     assert.equal(geometry.type, 'CylinderGeometry');
     assert.equal(geometry.parameters.radiusTop, 1);
@@ -202,10 +173,23 @@ suite('standard geometries', function () {
     assert.equal(geometry.parameters.thetaLength, degToRad(350));
   });
 
+  test('icosahedron', function () {
+    var el = this.el;
+    var geometry;
+    el.setAttribute('geometry', {
+      buffer: false, primitive: 'icosahedron', detail: 0, radius: 5});
+
+    geometry = el.getObject3D('mesh').geometry;
+    assert.equal(geometry.type, 'IcosahedronGeometry');
+    assert.equal(geometry.parameters.radius, 5);
+    assert.equal(geometry.parameters.detail, 0);
+  });
+
   test('plane', function () {
     var el = this.el;
     var geometry;
     el.setAttribute('geometry', {buffer: false, primitive: 'plane', width: 1, height: 2});
+
     geometry = el.getObject3D('mesh').geometry;
     assert.equal(geometry.type, 'PlaneGeometry');
     assert.equal(geometry.parameters.width, 1);
@@ -217,6 +201,7 @@ suite('standard geometries', function () {
     var geometry;
     el.setAttribute('geometry', {
       buffer: false, primitive: 'ring', radiusInner: 1, radiusOuter: 2, segmentsTheta: 3});
+
     geometry = el.getObject3D('mesh').geometry;
     assert.equal(geometry.type, 'RingGeometry');
     assert.equal(geometry.parameters.innerRadius, 1);
@@ -231,6 +216,7 @@ suite('standard geometries', function () {
       buffer: false, primitive: 'sphere', radius: 1, segmentsWidth: 2, segmentsHeight: 3,
       phiStart: 45, phiLength: 90, thetaStart: 45
     });
+
     geometry = el.getObject3D('mesh').geometry;
     assert.equal(geometry.type, 'SphereGeometry');
     assert.equal(geometry.parameters.radius, 1);
@@ -249,6 +235,7 @@ suite('standard geometries', function () {
       buffer: false, primitive: 'torus', radius: 1, radiusTubular: 2, segmentsRadial: 3,
       segmentsTubular: 4, arc: 350
     });
+
     geometry = el.getObject3D('mesh').geometry;
     assert.equal(geometry.type, 'TorusGeometry');
     assert.equal(geometry.parameters.radius, 1);
@@ -265,6 +252,7 @@ suite('standard geometries', function () {
       buffer: false, primitive: 'torusKnot', radius: 1, radiusTubular: 2, segmentsRadial: 3,
       segmentsTubular: 4, p: 5, q: 6
     });
+
     geometry = el.getObject3D('mesh').geometry;
     assert.equal(geometry.type, 'TorusKnotGeometry');
     assert.equal(geometry.parameters.radius, 1);